add support for RLE block types
Hand-crafted a zst test case, since I could not get the zstd tool to
compress with it.

Funnily enough, the result can be decoded by zstd's educational decoder
and mine, but not by the main zstd tool. Perhaps I got the xxhash wrong.
For now, being in line with the educational decoder is better than not
having an implemetation at all.