support overlapping offset copies
If we've only written 50 bytes, and we should write 100 from an offset
of 50, that means we should end up with the first 50 bytes repeated
three times.

However, we cannot simply trust the builtin copy to do this. Since it
deals with overlapping source and destination slices, we will end up
with the initial 50 bytes repeated twice, then 50 null characters. Chunk
the copies to make sure that doesn't happen.