update to the latest wuffs release
"!??" was replaced with "?", and "= try" with "=?". And effects within
subexpressions are banned. These are easy changes to apply.

However, declarations must now go at the top of each function. This
requires quite a bit of manual labor.

Finally, fix up the build script as the release snapshot is now
elsewhere.